requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.9.24
charset-normalizer==3.0.1
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
dnspython==2.3.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.9.0
pip==22.2.2
platformdirs==2.5.2
pymongo==4.3.3
pytz==2022.6
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
setuptools==65.5.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
wheel==0.37.1
